samples/08-hybrid-model-routing/requirements.txt
# Sample requirements - using ToolWeaver from PyPI
toolweaver[small-model]==0.1.3

# Environment variables
python-dotenv>=1.0.0
